+++ python/lsst/ts/wep/task/calcZernikesUnpairedTask.py
@@ -27,7 +27,7 @@ __all__ = [
 
 import lsst.pipe.base as pipeBase
 import numpy as np
-from astropy.table import QTable
+from astropy.table import QTable, Table
 from lsst.pipe.base import connectionTypes
 from lsst.ts.wep.task.calcZernikesTask import CalcZernikesTask, CalcZernikesTaskConfig
 from lsst.ts.wep.task.donutStamps import DonutStamps
@@ -36,13 +36,19 @@ from lsst.utils.timer import timeMethod
 
 class CalcZernikesUnpairedTaskConnections(
     pipeBase.PipelineTaskConnections,
-    dimensions=("visit", "detector", "instrument"),  # type: ignore
+    dimensions=("visit", "detector", "instrument", "physical_filter"),  # type: ignore
 ):
     donutStamps = connectionTypes.Input(
         doc="Defocused Donut Postage Stamp Images",
         dimensions=("visit", "detector", "instrument"),
         storageClass="StampsBase",
         name="donutStamps",
+    )
+    intrinsicTable = connectionTypes.Input(
+        doc="Intrinsic Zernike Map for the instrument",
+        dimensions=("detector", "instrument", "physical_filter"),
+        storageClass="ArrowAstropy",
+        name="intrinsic_aberrations_temp",
     )
     outputZernikesRaw = connectionTypes.Output(
         doc="Zernike Coefficients from all donuts",
@@ -87,6 +93,7 @@ class CalcZernikesUnpairedTask(CalcZernikesTask):
     def run(
         self,
         donutStamps: DonutStamps,
+        intrinsicTable: Table,
         numCores: int = 1,
     ) -> pipeBase.Struct:
         if len(donutStamps) == 0:
@@ -110,25 +117,42 @@ class CalcZernikesUnpairedTask(CalcZernikesTask):
             selectedDonuts = donutStamps
             donutQualityTable = QTable([])
 
-        # Assign stamps to either intra or extra
+        # Assign stamps to either intra or extra, and build intrinsic map
         defocalType = donutStamps.metadata["DFC_TYPE"]
         if defocalType == "extra":
+            self.stampsIntra = DonutStamps([])
             self.stampsExtra = selectedDonuts
             if len(donutQualityTable) > 0:
                 donutQualityTable["DEFOCAL_TYPE"] = "extra"
+            self.intrinsicMapExtra = self._createIntrinsicMap(intrinsicTable)
+            self.intrinsicMapIntra = None
         else:
             self.stampsIntra = selectedDonuts
+            self.stampsExtra = DonutStamps([])
             if len(donutQualityTable) > 0:
                 donutQualityTable["DEFOCAL_TYPE"] = "intra"
-        zkCoeffRaw = self.estimateZernikes.run(self.stampsExtra, self.stampsIntra, numCores=numCores)
-        # Combine Zernikes
-        zkCoeffCombined = self.combineZernikes.run(zkCoeffRaw.zernikes)
+            self.intrinsicMapExtra = None
+            self.intrinsicMapIntra = self._createIntrinsicMap(intrinsicTable)
 
-        zkTable = self.createZkTable(zkCoeffRaw, zkCoeffCombined)
+        # Estimate Zernikes
+        zkCoeffRaw = self.estimateZernikes.run(
+            self.stampsExtra,
+            self.stampsIntra,
+            numCores=numCores,
+        )
+
+        # Save the outputs in the table
+        zkTable = self.createZkTable(zkCoeffRaw)
         zkTable.meta["estimatorInfo"] = zkCoeffRaw.wfEstInfo
 
+        # Combine Zernikes
+        zkTable = self.combineZernikes.run(zkTable).combinedTable
+
+        avg = zkTable[zkTable["label"] == "average"]
+        outputZernikesAvg = np.array([avg[col].to_value("um")[0] for col in avg.meta["opd_columns"]])
+
         return pipeBase.Struct(
-            outputZernikesAvg=np.atleast_2d(np.array(zkCoeffCombined.combinedZernikes)),
+            outputZernikesAvg=np.atleast_2d(np.array(outputZernikesAvg)),
             outputZernikesRaw=np.atleast_2d(np.array(zkCoeffRaw.zernikes)),
             zernikes=zkTable,
             donutQualityTable=donutQualityTable,