+++ tests/task/test_calcZernikesTieTaskCwfs.py
@@ -67,7 +67,7 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
                 cleanUpCmd = writeCleanUpRepoCmd(cls.repoDir, cls.runName)
                 runProgram(cleanUpCmd)
 
-            collections = "refcats/gen2,LSSTCam/calib,LSSTCam/raw/all"
+            collections = "refcats/gen2,LSSTCam/calib,LSSTCam/raw/all,LSSTCam/aos/intrinsic"
             instrument = "lsst.obs.lsst.LsstCam"
             pipelineYaml = os.path.join(testPipelineConfigDir, "testCalcZernikesCwfsSetupPipeline.yaml")
 
@@ -99,12 +99,14 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
             "detector": 191,
             "exposure": 4021123106000,
             "visit": 4021123106000,
+            "physical_filter": "g",
         }
         self.dataIdIntra = {
             "instrument": "LSSTCam",
             "detector": 191,
             "exposure": 4021123106000,
             "visit": 4021123106000,
+            "physical_filter": "g",
         }
         self.donutStampsExtra = self.butler.get(
             "donutStampsExtra", dataId=self.dataIdExtra, collections=[self.runName]
@@ -112,6 +114,18 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
         self.donutStampsIntra = self.butler.get(
             "donutStampsIntra", dataId=self.dataIdExtra, collections=[self.runName]
         )
+        self.intrinsicTables = [
+            self.butler.get(
+                "intrinsic_aberrations_temp",
+                dataId=self.dataIdExtra,
+                collections=["LSSTCam/aos/intrinsic"],
+            ),
+            self.butler.get(
+                "intrinsic_aberrations_temp",
+                dataId=self.dataIdIntra | {"detector": 192},
+                collections=["LSSTCam/aos/intrinsic"],
+            ),
+        ]
 
     def testValidateConfigs(self) -> None:
         self.assertEqual(type(self.task.combineZernikes), CombineZernikesSigmaClipTask)
@@ -141,7 +155,7 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
         donutStampsExtra = DonutStamps.readFits(os.path.join(donutStampDir, "R04_SW0_donutStamps.fits"))
         donutStampsIntra = DonutStamps.readFits(os.path.join(donutStampDir, "R04_SW1_donutStamps.fits"))
         zernCoeffAllR04 = self.task.estimateZernikes.run(donutStampsExtra, donutStampsIntra).zernikes
-        zernCoeffAvgR04 = self.task.combineZernikes.run(zernCoeffAllR04).combinedZernikes
+        zernCoeffAvgR04 = zernCoeffAllR04.mean(axis=0)
         trueZernCoeffR04 = np.array(
             [
                 -3.98568849e-01,
@@ -179,7 +193,7 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
         donutStampsExtra = DonutStamps.readFits(os.path.join(donutStampDir, "R40_SW0_donutStamps.fits"))
         donutStampsIntra = DonutStamps.readFits(os.path.join(donutStampDir, "R40_SW1_donutStamps.fits"))
         zernCoeffAllR40 = self.task.estimateZernikes.run(donutStampsExtra, donutStampsIntra).zernikes
-        zernCoeffAvgR40 = self.task.combineZernikes.run(zernCoeffAllR40).combinedZernikes
+        zernCoeffAvgR40 = zernCoeffAllR40.mean(axis=0)
         trueZernCoeffR40 = np.array(
             [
                 -4.51261752e-01,
@@ -213,13 +227,6 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
         # from the OPD truth as a sanity check
         self.assertLess(np.sqrt(np.sum(np.square(zernCoeffAvgR40 - trueZernCoeffR40))), 0.35)
 
-    def testGetCombinedZernikes(self) -> None:
-        testArr = np.zeros((2, 25))
-        testArr[1] += 2.0
-        combinedZernikesStruct = self.task.combineZernikes.run(testArr)
-        np.testing.assert_array_equal(combinedZernikesStruct.combinedZernikes, np.ones(25))
-        np.testing.assert_array_equal(combinedZernikesStruct.flags, np.zeros(len(testArr)))
-
     def testWithAndWithoutPairs(self) -> None:
         # Load the test data
         donutStampDir = os.path.join(self.testDataDir, "donutImg", "donutStamps")
@@ -228,27 +235,22 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
 
         # First estimate without pairs
         zkAllExtra = self.task.estimateZernikes.run(donutStampsExtra, []).zernikes
-        zkAvgExtra = self.task.combineZernikes.run(zkAllExtra).combinedZernikes
         zkAllIntra = self.task.estimateZernikes.run([], donutStampsIntra).zernikes
-        zkAvgIntra = self.task.combineZernikes.run(zkAllIntra).combinedZernikes
 
         # Now estimate with pairs
         zkAllPairs = self.task.estimateZernikes.run(donutStampsExtra, donutStampsIntra).zernikes
-        zkAvgPairs = self.task.combineZernikes.run(zkAllPairs).combinedZernikes
 
         # Check that all have same number of Zernike coeffs
         self.assertEqual(zkAllExtra.shape[1], zkAllPairs.shape[1])
         self.assertEqual(zkAllIntra.shape[1], zkAllPairs.shape[1])
-        self.assertEqual(len(zkAvgExtra), len(zkAvgPairs))
-        self.assertEqual(len(zkAvgIntra), len(zkAvgPairs))
 
         # Check that unpaired is at least as long as paired
         self.assertGreaterEqual(zkAllExtra.shape[0], zkAllPairs.shape[0])
         self.assertGreaterEqual(zkAllIntra.shape[0], zkAllPairs.shape[0])
 
         # Check that the averages are similar
-        zkAvgUnpaired = np.mean([zkAvgExtra, zkAvgIntra], axis=0)
-        self.assertLess(np.sqrt(np.sum(np.square(zkAvgPairs - zkAvgUnpaired))), 0.30)
+        zkAvgUnpaired = np.mean([zkAllExtra.mean(axis=0), zkAllIntra.mean(axis=0)], axis=0)
+        self.assertLess(np.sqrt(np.sum(np.square(zkAllPairs.mean(axis=0) - zkAvgUnpaired))), 0.33)
 
     def testUnevenPairs(self) -> None:
         # Test for when you have more of either extra or intra
@@ -260,7 +262,7 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
         stampsExtra.extend([stampsExtra[0]])
 
         # Now estimate Zernikes
-        self.task.run(stampsExtra, stampsIntra)
+        self.task.run(stampsExtra, stampsIntra, self.intrinsicTables)
 
     def testRequireConverge(self) -> None:
         config = CalcZernikesTaskConfig()
@@ -285,12 +287,15 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
         donutStampsIntra = DonutStamps.readFits(os.path.join(donutStampDir, "R04_SW1_donutStamps.fits"))
 
         # Estimate Zernikes 4, 5, 6
-        self.task.config.estimateZernikes.nollIndices = [4, 5, 6]
-        zk0 = self.task.estimateZernikes.run(donutStampsExtra, donutStampsIntra).zernikes[0]
+        config = CalcZernikesTaskConfig()
+        config.estimateZernikes.nollIndices = [4, 5, 6]
+        task = CalcZernikesTask(config=config, name="Test Noll Indices 1")
+        zk0 = task.estimateZernikes.run(donutStampsExtra, donutStampsIntra).zernikes[0]
 
         # Estimate Zernikes 4, 5, 6, 20, 21
-        self.task.config.estimateZernikes.nollIndices = [4, 5, 6, 20, 21]
-        zk1 = self.task.estimateZernikes.run(donutStampsExtra, donutStampsIntra).zernikes[0]
+        config.estimateZernikes.nollIndices = [4, 5, 6, 20, 21]
+        task = CalcZernikesTask(config=config, name="Test Noll Indices 2")
+        zk1 = task.estimateZernikes.run(donutStampsExtra, donutStampsIntra).zernikes[0]
 
         # Check lengths
         self.assertEqual(len(zk0), 3)
@@ -303,8 +308,8 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
     def testTableMetadata(self) -> None:
         # First estimate without pairs
         emptyStamps = DonutStamps([], metadata=self.donutStampsExtra.metadata)
-        zkCalcExtra = self.task.run(self.donutStampsExtra, emptyStamps).zernikes
-        zkCalcIntra = self.task.run(emptyStamps, self.donutStampsIntra).zernikes
+        zkCalcExtra = self.task.run(self.donutStampsExtra, emptyStamps, self.intrinsicTables).zernikes
+        zkCalcIntra = self.task.run(emptyStamps, self.donutStampsIntra, self.intrinsicTables).zernikes
 
         # Check metadata keys exist for extra case
         self.assertIn("cam_name", zkCalcExtra.meta)
@@ -327,7 +332,9 @@ class TestCalcZernikesTieTaskCwfs(lsst.utils.tests.TestCase):
             self.assertEqual(dict_["mjd"], self.donutStampsIntra.metadata["MJD"])
 
         # Now estimate with pairs
-        zkCalcPairs = self.task.run(self.donutStampsExtra, self.donutStampsIntra).zernikes
+        zkCalcPairs = self.task.run(
+            self.donutStampsExtra, self.donutStampsIntra, self.intrinsicTables
+        ).zernikes
 
         # Check metadata keys exist for pairs case
         self.assertIn("cam_name", zkCalcPairs.meta)